moved functions to their respective module
