+++ __main__.py
@@ -2,8 +2,8 @@ import argparse
 import logging
 import os
 
-import converter
-import download
+from download import download_manga
+from converter import convert
 import search
 from config import JPG_DIR, MANGA_DIR
 from menu import SearchMenu
@@ -16,31 +16,6 @@ logging.basicConfig(
 
 
 logger = logging.getLogger(__name__)
-
-
-def download_manga(manga, volume):
-    downloader = download.DownloadManga(manga)
-    if not os.path.exists(JPG_DIR):
-        os.makedirs(JPG_DIR)
-    if volume.isdigit() or isinstance(volume, int):
-        downloader.download_volume(volume)
-    elif isinstance(volume, str):
-        downloader.download_volumes(volume)
-    else:
-        downloader.download_all_volumes()
-
-
-def convert(manga, volume, cbz):
-    conversion = converter.Conversion(manga)
-    conversion.type = "cbz" if cbz else "pdf"
-    if not os.path.exists(MANGA_DIR):
-        os.makedirs(MANGA_DIR)
-    if volume.isdigit() or isinstance(volume, int):
-        conversion.convert_volume(volume)
-    elif isinstance(volume, str):
-        conversion.convert_volumes(volume)
-    else:
-        conversion.convert_all_volumes()
 
 
 def cli():
